requirements.txt
faiss-cpu==1.8.0
sentence-transformers==2.5.1
numpy==1.26.4
sqlalchemy==2.0.28
torch>=2.0.0
transformers>=4.36.0
huggingface-hub>=0.20.0